chore: provider directory 안으로 provider 관련 코드 이동
