+++ apps/commerce-client-web/src/lib/msw-api.ts
@@ -1,0 +1,12 @@
+import ky from 'ky';
+
+const mswApi = ky.create({
+  prefixUrl: 'http://localhost:3000/api/',
+  timeout: 10000,
+  retry: {
+    limit: 2,
+    statusCodes: [401],
+  },
+});
+
+export default mswApi;
