+++ apps/commerce-client-web/src/provider/app-provider.tsx
@@ -1,0 +1,22 @@
+'use client';
+
+import { PropsWithChildren } from 'react';
+import { QueryClientProvider } from '@tanstack/react-query';
+import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
+import getQueryClient from '@/configs/tanstack-query/get-query-client';
+import { SessionProvider } from 'next-auth/react';
+
+const AppProvider = ({ children }: PropsWithChildren) => {
+  const queryClient = getQueryClient();
+
+  return (
+    <SessionProvider>
+      <QueryClientProvider client={queryClient}>
+        {children}
+        <ReactQueryDevtools initialIsOpen={false} />
+      </QueryClientProvider>
+    </SessionProvider>
+  );
+};
+
+export default AppProvider;
